Problem: I do not like how the code is organized
